Bug fixing in progress.

+++ Evis.VMS/Evis.VMS.UI/Controllers/AccountController.cs
@@ -322,24 +322,27 @@ namespace Evis.VMS.UI.Controllers
                 {
                     file = Request.Files[fileName];
                     fName = string.Format("{0}_{1}", fileName, file.FileName);
-                    if (file != null && file.ContentLength > 0)
+                    if (!fName.Contains("file_"))
                     {
-                        var directoryPath = string.Format("{0}images\\UserImages", Server.MapPath(@"\"));
-                        if (!Directory.Exists(directoryPath))
+                        if (file != null && file.ContentLength > 0)
                         {
-                            Directory.CreateDirectory(directoryPath);
-                        }
-                        var originalDirectory = new DirectoryInfo(directoryPath);
-                        fileWithPath = System.IO.Path.Combine(originalDirectory.ToString(), fName);
-                        var fileName1 = Path.GetFileName(fName);
-                        var isExists = System.IO.File.Exists(fileWithPath);
+                            var directoryPath = string.Format("{0}images\\UserImages", Server.MapPath(@"\"));
+                            if (!Directory.Exists(directoryPath))
+                            {
+                                Directory.CreateDirectory(directoryPath);
+                            }
+                            var originalDirectory = new DirectoryInfo(directoryPath);
+                            fileWithPath = System.IO.Path.Combine(originalDirectory.ToString(), fName);
+                            var fileName1 = Path.GetFileName(fName);
+                            var isExists = System.IO.File.Exists(fileWithPath);
 
-                        if (isExists)
-                        {
-                            System.IO.File.Delete(fileWithPath);
-                        }
+                            if (isExists)
+                            {
+                                System.IO.File.Delete(fileWithPath);
+                            }
 
-                        file.SaveAs(fileWithPath);
+                            file.SaveAs(fileWithPath);
+                        }
                     }
                 }
             }
